+++ src/server-actions/Channel/Channel.ts
@@ -33,6 +33,7 @@ import { GetUserPermissionsParsedAction } from "../UserRoles/UserRole"
 import { PermissionChecker } from "@/src/lib/PermissionCheker"
 import {
   attachCommunityUser,
+  ensureCommunityMembership,
   getCommunityUsers
 } from "@/src/db/data-access/communities/query"
 
@@ -199,23 +200,7 @@ export const AttachChannelUserAction = CreateServerAction(
         attachUserRole?.viewerRole?.name
       )
       if (channel?.community_id) {
-        const communityMembers = await getCommunityUsers(channel?.community_id)
-        const communityUserIds = communityMembers.map((cu) => cu.user_id)
-
-        const isMemberCommunity = communityUserIds.includes(userId)
-
-        if (!isMemberCommunity) {
-          const attachCommunityUserRole = await getAndAssignViewerRoles(
-            userId,
-            "community_viewer",
-            channel.community_id
-          )
-          await attachCommunityUser(
-            channel.community_id,
-            userId,
-            attachCommunityUserRole?.viewerRole?.name
-          )
-        }
+        await ensureCommunityMembership(channel.community_id, userId)
       }
       return { success: true, data: channelUser }
     } catch (error) {
